[quant] Create default custom modules for LSTM and MHA
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71192

Currently we expect the users to provide custom modules for LSTM and MHA. However, as we almost always ask the users to use those modules in the custom context, it is better to make this behavior default. In this case we try to align with the base quantization API, if the user specifies a custom_config_dict then that is used, however if the value is left as None then the default is used. If a user would like to both use the default and modify it, they have to do so manually, however the default is accessible by get_default_custom_config_dict

Additionally, the NS which uses prepare to insert custom observers for
its purposes had to be slightly modified to pass in an empty
custom_config_dict in order to avoid modifying the custom modules.

Differential Revision: [D33452890](https://our.internmc.facebook.com/intern/diff/D33452890/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33452890/)!
ghstack-source-id: a8326adcc9677eeeb6f04005fbda13a637a839e5