[quant] Reordering the imports in the torch/__init__.py
Because the AO stuff depends on the torch packages, but very few (if any)
torch packages depend on AO, we are moving the imports lower.
That will reduce the probability of cyclic imports, as by the time the
AO would start importing, the rest of the torch would be already imported.

ghstack-source-id: 5e73055fa2294e0457237229eecc5d24a061ecba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77065